Switch to using multinomial distributions

+++ generator/util.py
@@ -2,7 +2,7 @@ import numpy as np
 
 
 def generate_expression_profile(cell_spec, n_genes, cell_type, rng: np.random.Generator):
-    gene_p = np.ndarray(shape=n_genes, dtype="float32")
+    gene_p = np.ndarray(shape=n_genes, dtype="float64")
     weights = cell_spec.X[cell_type]
     weights_std = cell_spec.layers["std"][cell_type]
     gene_p = np.maximum(rng.normal(weights, weights_std), 0)
